put into a separete file the VUE.directives
